pyproject.toml
[tool.poetry]
name = "imageflipper"
version = "0.2.0"
description = ""
authors = ["elferia <a.t.qu.qu.qu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.11.1"
autopep8 = "^1.5"
rope = "^0.16.0"

[tool.poetry.scripts]
imageflipper = 'imageflipper.main:run'
overresizer = 'imageflipper.overresize:run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
